Structure IW scanning into ap-scanner and add error handling
This commit now adds more modularity to running commands so hopefully it
is easier to expand to commands on other systems. Also, it adds error
handling for different errors such as a failed scan, incorrect
permissions, etc.

The cmd.go file is responsible for all the functions related to running
the 'iw' command. This will probably need to be further split in the
future.

The errors.go file contains common custom errors in the program. Once
the sender is implemented, this file will probably need to be in a
separate directory, such as `common/`.