Add a unit test for parsing the scan output of the IW command
This was the only functionality I found I could realistically test
without relying on the internal state of the machine (commands being
ran, interfaces being found). It doesn't appear to be easy (or even
possible) to mock many calls in Go without refactoring the code to make
it use global variables for funtions or custom parameters for the specific
function to use (which in my opinion makes it a lot more unreadable).

I made the testify package required as that will be very helpful when
writing tests (asserting, requiring, mocking).