Add GitHub workflows to test code after pushes and PRs
Four different CI tests will be run: build, test, lint, and format
Additionally, codespell is run to check any comments or documentation
for typos.

The lint action will only run golangci-lint, not every linter in `make
lint`. Additionally, the format step will only check if files need to be
formatted, but `go mod tidy` will not be run.

The steps could be split into different jobs for a faster runtime, but
each step needs to be run inside the wardriver path, so I decided to
group them together. Also, the little more amount of time it takes to
run doesn't matter much (we can't take advantage of GitHub running jobs
in parallel) as it is a small project right now.