Send all access points every scan interval
This significantly reduces the number of requests being made. There are
currently a couple of TODOs that need to be done. Specifically, only the
APs that failed to be inserted into the DB from the server should be
enqueued, not all of them. This will need some change/improvement on the
server side as well